Update and rename executeCode.py to execute_code.py
